TypeSignature to util file
