+++ libs/mock-builder/src/location.rs
@@ -1,5 +1,7 @@
 use frame_support::StorageHasher;
 
+use super::util::TypeSignature;
+
 /// Absolute string identification of function.
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct FunctionLocation(String);
@@ -60,12 +62,7 @@ impl FunctionLocation {
 
 	/// Add a representation of the function input and output types
 	pub fn append_type_signature<I, O>(self) -> Self {
-		Self(format!(
-			"{}:{}->{}",
-			self.0,
-			std::any::type_name::<I>(),
-			std::any::type_name::<O>(),
-		))
+		Self(format!("{}:{}", self.0, TypeSignature::new::<I, O>()))
 	}
 
 	/// Generate a hash of the location
